Tweak benchmarks
