+++ benchmarks/Serialization/ResourceSerializationBenchmarks.cs
@@ -12,6 +12,7 @@ using JsonApiDotNetCore.Serialization.Objects;
 namespace Benchmarks.Serialization;
 
 [MarkdownExporter]
+[MemoryDiagnoser]
 // ReSharper disable once ClassCanBeSealed.Global
 public class ResourceSerializationBenchmarks : SerializationBenchmarkBase
 {
